// src/app/page/master/master-pegawai-sk-hitung-pajak/master-pegawai-sk-hitung-pajak.interface.ts
import {EventEmitter} from '@angular/core';

export interface MasterPegawaiSkHitungPajak {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
 }